% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IB.R
\name{IB}
\alias{IB}
\title{IBREST Class}
\description{
Get data data from IB Client Portal API.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{host}}{Host, by default localhost}

\item{\code{port}}{Port, by default 5000}

\item{\code{base_url}}{Base url}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IB-new}{\code{IB$new()}}
\item \href{#method-IB-get}{\code{IB$get()}}
\item \href{#method-IB-post}{\code{IB$post()}}
\item \href{#method-IB-delete}{\code{IB$delete()}}
\item \href{#method-IB-get_conids_by_exchange}{\code{IB$get_conids_by_exchange()}}
\item \href{#method-IB-search_contract_by_symbol}{\code{IB$search_contract_by_symbol()}}
\item \href{#method-IB-get_sec_definfo}{\code{IB$get_sec_definfo()}}
\item \href{#method-IB-get_trading_schedule}{\code{IB$get_trading_schedule()}}
\item \href{#method-IB-get_position_by_conid}{\code{IB$get_position_by_conid()}}
\item \href{#method-IB-place_order}{\code{IB$place_order()}}
\item \href{#method-IB-confirm_order}{\code{IB$confirm_order()}}
\item \href{#method-IB-get_order_status}{\code{IB$get_order_status()}}
\item \href{#method-IB-place_and_confirm_order}{\code{IB$place_and_confirm_order()}}
\item \href{#method-IB-get_historical_data_beta}{\code{IB$get_historical_data_beta()}}
\item \href{#method-IB-get_historical_data}{\code{IB$get_historical_data()}}
\item \href{#method-IB-cancel_order}{\code{IB$cancel_order()}}
\item \href{#method-IB-get_trades}{\code{IB$get_trades()}}
\item \href{#method-IB-get_portfolio_accounts}{\code{IB$get_portfolio_accounts()}}
\item \href{#method-IB-get_portfolio_subaccounts}{\code{IB$get_portfolio_subaccounts()}}
\item \href{#method-IB-get_portfolio_summary}{\code{IB$get_portfolio_summary()}}
\item \href{#method-IB-clone}{\code{IB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-new"></a>}}
\if{latex}{\out{\hypertarget{method-IB-new}{}}}
\subsection{Method \code{new()}}{
Create a new IB object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$new(host = "localhost", port = 5000L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{host}}{Host, by default localhost}

\item{\code{port}}{Port, by default 5000}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `IB` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get}{}}}
\subsection{Method \code{get()}}{
Performs a GET request to the specified Interactive Brokers API endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get(endpoint = "/sso/validate", query = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The endpoint path for the GET request.}

\item{\code{query}}{A list of query parameters for the GET request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the GET request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-post"></a>}}
\if{latex}{\out{\hypertarget{method-IB-post}{}}}
\subsection{Method \code{post()}}{
Performs a POST request to the specified Interactive Brokers API endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$post(endpoint = "/tickle", body = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The endpoint path for the POST request.}

\item{\code{body}}{The body of the POST request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the POST request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-delete"></a>}}
\if{latex}{\out{\hypertarget{method-IB-delete}{}}}
\subsection{Method \code{delete()}}{
Performs a DELETE request to the specified Interactive Brokers API endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$delete(endpoint)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{endpoint}}{The endpoint path for the DELETE request.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the DELETE request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_conids_by_exchange"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_conids_by_exchange}{}}}
\subsection{Method \code{get_conids_by_exchange()}}{
Retrieves all contracts available on a specified exchange.
Note: This is only available for Stock contracts.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_conids_by_exchange(exchange)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exchange}}{A string specifying the exchange.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with columns.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-search_contract_by_symbol"></a>}}
\if{latex}{\out{\hypertarget{method-IB-search_contract_by_symbol}{}}}
\subsection{Method \code{search_contract_by_symbol()}}{
Searches by underlying symbol or company name and relays back what derivative contract(s) it has.
This endpoint must be called before using /secdef/info.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$search_contract_by_symbol(symbol, name = NULL, secType = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{symbol}}{String, required, underlying symbol of interest or company name if ‘name’ is set to true.}

\item{\code{name}}{Boolean, optional, determines if symbol reflects company name or ticker symbol.}

\item{\code{secType}}{String, optional, declares underlying security type.
Valid values: "STK", "IND", "BOND".}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the contract such as conid, companyName, symbol, etc.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_sec_definfo"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_sec_definfo}{}}}
\subsection{Method \code{get_sec_definfo()}}{
Retrieves security definition information for a given contract identifier (conid).
This method is applicable for various security types including futures, options, warrants, cash, and CFDs.
For derivatives, /iserver/secdef/search must be called first.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_sec_definfo(
  conid,
  sectype,
  month = NULL,
  exchange = NULL,
  strike = NULL,
  right = NULL,
  issuerId = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conid}}{String, required, contract identifier of the underlying or the final derivative conid.}

\item{\code{sectype}}{String, required, security type of the requested contract.}

\item{\code{month}}{String, required for derivatives, expiration month for the given derivative.}

\item{\code{exchange}}{String, optional, exchange to receive information for in relation to the contract.}

\item{\code{strike}}{String, required for options and futures options, strike price for the requested contract.}

\item{\code{right}}{String, required for options, "C" for Call or "P" for Put.}

\item{\code{issuerId}}{String, required for bonds, issuerId for the given bond issuer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the contract such as conid, ticker, secType, etc.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_trading_schedule"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_trading_schedule}{}}}
\subsection{Method \code{get_trading_schedule()}}{
Retrieves the trading schedule up to a month for the requested contract.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_trading_schedule(
  assetClass,
  symbol,
  exchange = NULL,
  exchangeFilter = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assetClass}}{String, required, security type of the given contract.
Possible values: STK, OPT, FUT, CFD, WAR, SWP, FND, BND, ICS.}

\item{\code{symbol}}{String, required, symbol for the contract.}

\item{\code{exchange}}{String, optional, primary exchange of the contract.}

\item{\code{exchangeFilter}}{String, optional, exchanges to retrieve data from.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the trading schedule details.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_position_by_conid"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_position_by_conid}{}}}
\subsection{Method \code{get_position_by_conid()}}{
Retrieves position details for the specified contract identifier (conid).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_position_by_conid(accountId, conId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accountId}}{String, required, the account ID for which the position should be retrieved.}

\item{\code{conId}}{String, required, the contract ID to receive position information on.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing position details for the specified conid.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-place_order"></a>}}
\if{latex}{\out{\hypertarget{method-IB-place_order}{}}}
\subsection{Method \code{place_order()}}{
Submits orders when connected to an IServer Brokerage Session.
Supports various advanced order types and additional details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$place_order(accountID, orders)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accountID}}{String, required, the account ID for which the order should be placed.}

\item{\code{orders}}{Array of Objects, required, used to specify the order content.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the order status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-confirm_order"></a>}}
\if{latex}{\out{\hypertarget{method-IB-confirm_order}{}}}
\subsection{Method \code{confirm_order()}}{
Confirms the order reply, agreeing or declining the message to transmit or discard the order.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$confirm_order(replyId, confirmed = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{replyId}}{String, required, the ID value from the prior order request.}

\item{\code{confirmed}}{Boolean, required, pass TRUE to agree to the message and transmit the order, or FALSE to decline and discard the order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the order confirmation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_order_status"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_order_status}{}}}
\subsection{Method \code{get_order_status()}}{
Retrieves the status of an individual order using the orderId.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_order_status(orderId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{orderId}}{String, required, order identifier for the placed order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the order status.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-place_and_confirm_order"></a>}}
\if{latex}{\out{\hypertarget{method-IB-place_and_confirm_order}{}}}
\subsection{Method \code{place_and_confirm_order()}}{
Places an order and handles the confirmation process.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$place_and_confirm_order(accountID, orders)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accountID}}{String, required, the account ID for which the order should be placed.}

\item{\code{orders}}{Array of Objects, required, used to specify the order content.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the order placement and confirmation.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_historical_data_beta"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_historical_data_beta}{}}}
\subsection{Method \code{get_historical_data_beta()}}{
Retrieves a list of historical market data for a given contract identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_historical_data_beta(
  conid,
  period,
  bar,
  outsideRth = NULL,
  startTime = NULL,
  direction = NULL,
  barType = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conid}}{String, required, contract identifier for which data should be requested.}

\item{\code{period}}{String, required, duration for which data should be requested.}

\item{\code{bar}}{String, required, bar size for which bars should be returned.}

\item{\code{outsideRth}}{Boolean, optional, define if data should be returned for trades outside regular trading hours.}

\item{\code{startTime}}{String, optional, specify the value from where historical data should be taken.}

\item{\code{direction}}{String, optional, specify the direction from which market data should be returned.}

\item{\code{barType}}{String, optional, returns valid bar types for which data may be requested.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing historical market data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_historical_data"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_historical_data}{}}}
\subsection{Method \code{get_historical_data()}}{
Retrieves historical market data for a given contract identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_historical_data(
  conid,
  period,
  bar,
  exchange = NULL,
  startTime = NULL,
  outsideRth = NULL,
  clean = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conid}}{String, required, contract identifier for the ticker symbol of interest.}

\item{\code{period}}{String, required, overall duration for which data should be returned.}

\item{\code{bar}}{String, required, individual bars of data to be returned.}

\item{\code{exchange}}{String, optional, exchange to receive data from.}

\item{\code{startTime}}{String, optional, starting date of the request duration.}

\item{\code{outsideRth}}{Boolean, optional, determine if data after regular trading hours is needed.}

\item{\code{clean}}{Boolean, optional, if TRUE extract data object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing historical market data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-cancel_order"></a>}}
\if{latex}{\out{\hypertarget{method-IB-cancel_order}{}}}
\subsection{Method \code{cancel_order()}}{
Cancels an open order for the specified account and order ID.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$cancel_order(accountId, orderId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accountId}}{String, required, the account ID for which the order should be canceled.}

\item{\code{orderId}}{String, required, the order ID that should be canceled.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing the confirmation of the cancellation request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_trades"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_trades}{}}}
\subsection{Method \code{get_trades()}}{
Retrieves a list of trades for the currently selected account for the current day and six previous days.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_trades(days = NULL, accountId = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{days}}{String, optional, specify the number of days to receive executions for, up to a maximum of 7 days.}

\item{\code{accountId}}{String, optional, include a specific account identifier or allocation group to retrieve trades for.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of trades.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_portfolio_accounts"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_portfolio_accounts}{}}}
\subsection{Method \code{get_portfolio_accounts()}}{
Retrieves a list of accounts for which the user can view position and account information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_portfolio_accounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing details of portfolio accounts.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_portfolio_subaccounts"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_portfolio_subaccounts}{}}}
\subsection{Method \code{get_portfolio_subaccounts()}}{
Retrieves a list of sub-accounts for which the user can view position and account-related information.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_portfolio_subaccounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing details of portfolio subaccounts.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-get_portfolio_summary"></a>}}
\if{latex}{\out{\hypertarget{method-IB-get_portfolio_summary}{}}}
\subsection{Method \code{get_portfolio_summary()}}{
Retrieves information regarding settled cash, cash balances, etc., in the account’s base currency and other cash balances held in other currencies.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$get_portfolio_summary(accountId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{accountId}}{String, required, specify the account ID for which you require ledger information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing details of the portfolio summary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
