% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{Utils}
\alias{Utils}
\title{Utils Class}
\description{
Utils method to work with IB data.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Utils-order_result_to_html}{\code{Utils$order_result_to_html()}}
\item \href{#method-Utils-clone}{\code{Utils$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Utils-order_result_to_html"></a>}}
\if{latex}{\out{\hypertarget{method-Utils-order_result_to_html}{}}}
\subsection{Method \code{order_result_to_html()}}{
Performs a GET request to the specified Interactive Brokers API endpoint.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Utils$order_result_to_html(order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{order}}{Filled order response}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The response from the GET request.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Utils-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Utils-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Utils$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
